- Update Curl to version number 7.5.13 - Remove some patches - Rewrite "get_ssl_handle" patch

+++ libs/curl/lib/parsedate.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: parsedate.c,v 1.19 2005/10/04 18:15:33 bagder Exp $
+ * $Id: parsedate.c,v 1.21 2005/12/30 00:35:21 curlvms Exp $
  ***************************************************************************/
 /*
   A brief summary of the date string formats this parser groks:
@@ -390,8 +390,8 @@ static time_t Curl_parsedate(const char *date)
   */
   t = mktime(&tm);
 
-  /* time zone adjust */
-  if(-1 != t) {
+  /* time zone adjust (cast t to int to compare to negative one) */
+  if(-1 != (int)t) {
     struct tm *gmt;
     long delta;
     time_t t2;
